Blogster Ai/requirements.txt
flask
bs4
pymongo
python-dotenv

```
